requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.7.4,<3.0.0

# LangGraph + LangChain (let pip resolve compatible versions together)
langgraph>=0.2.19,<0.3.0
langchain>=0.3.7,<0.4.0
langchain-core>=0.3.17,<0.4.0
langchain-groq>=0.2.1,<0.3.0
langchain-community>=0.3.7,<0.4.0

# LLM APIs
groq>=0.11.0

# Web Search
tavily-python>=0.5.0

# Arxiv
arxiv>=2.1.0

# Semantic Similarity (skip torch pin — pip picks compatible build)
sentence-transformers>=3.3.0
torch>=2.1.0

# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=14.0

# Gradio Frontend
gradio>=5.6.0,<6.0.0

# PDF generation
reportlab>=4.2.0
markdown2>=2.5.0

# Utilities
httpx>=0.27.0
aiofiles>=24.1.0
python-multipart>=0.0.17

# Mind Map & Visualization
matplotlib>=3.8.0
networkx>=3.2.0
